% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bickel_pal.R
\name{bickel_pal}
\alias{bickel_pal}
\title{Creates personal color palettes for plotting}
\usage{
bickel_pal(palette = "main", reverse = FALSE, ...)
}
\description{
Creates personal color palettes for plotting
}
\examples{
mypalette <- bickel_pal('cool')(7)
image(1:7,1,as.matrix(1:7),col=mypalette,xlab="Cool (sequential)",
ylab="",xaxt="n",yaxt="n",bty="n")

}
