% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papa_cols.R
\name{papa_cols}
\alias{papa_cols}
\title{Defines PJ color specs for palettes}
\usage{
papa_cols(...)
}
\description{
Defines PJ color specs for palettes
}
\examples{
mtcars \%>\%
ggplot(
    aes(
        x = mpg,
        y = cyl
   )
) +
geom_point(
   colour = papa_cols('red')
)

}
