% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papa_pal.R
\name{papa_pal}
\alias{papa_pal}
\title{Creates PJ color palettes for plotting}
\usage{
papa_pal(palette = "main", reverse = FALSE, ...)
}
\description{
Creates PJ color palettes for plotting
}
\examples{
mypalette <- papa_pal('cool')(7)
image(1:7,1,as.matrix(1:7),col=mypalette,xlab="Cool (sequential)",
ylab="",xaxt="n",yaxt="n",bty="n")

}
