% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_bickel.R
\name{scale_colour_bickel}
\alias{scale_colour_bickel}
\title{Define set of mappings from levels in the data aesthetic based on the bickel palette}
\usage{
scale_colour_bickel(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\description{
Define set of mappings from levels in the data aesthetic based on the bickel palette
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) +
geom_point(aes(colour = factor(cyl)))

p + scale_colour_bickel(palette = 'cool')

}
