% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_papa.R
\name{scale_fill_papa}
\alias{scale_fill_papa}
\title{Define set of mappings from levels in the data aesthetic based on the papa palette}
\usage{
scale_fill_papa(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\description{
Define set of mappings from levels in the data aesthetic based on the papa palette
}
\examples{
tibble(x = c('A', 'B', 'C'), y = runif(3, 10, 20)) \%>\%
ggplot(aes(x = x, y = y, fill = x)) + geom_col() + scale_fill_papa('hot')

}
